Prep for v0.3.0 release

+++ scripts/ci-e2e.sh
@@ -22,49 +22,53 @@
 set -o nounset
 set -o pipefail
 
-BOSKOS_HOST=${BOSKOS_HOST:-"boskos.test-pods.svc.cluster.local."}
+REPO_ROOT=$(dirname "${BASH_SOURCE[0]}")/..
+cd "${REPO_ROOT}" || exit 1
+
+# shellcheck source=../hack/ensure-go.sh
+source "${REPO_ROOT}/hack/ensure-go.sh"
+# shellcheck source=../hack/ensure-kind.sh
+source "${REPO_ROOT}/hack/ensure-kind.sh"
+# shellcheck source=../hack/ensure-kubectl.sh
+source "${REPO_ROOT}/hack/ensure-kubectl.sh"
+
 ARTIFACTS="${ARTIFACTS:-${PWD}/_artifacts}"
+mkdir -p "${ARTIFACTS}/logs/"
 
 # our exit handler (trap)
 cleanup() {
   # stop boskos heartbeat
   [[ -z ${HEART_BEAT_PID:-} ]] || kill -9 "${HEART_BEAT_PID}"
 }
-
 trap cleanup EXIT
 
-REPO_ROOT=$(dirname "${BASH_SOURCE[0]}")/..
-cd "${REPO_ROOT}" || exit 1
-
-echo "using boskos host to checkout project: ${BOSKOS_HOST}"
-
 # If BOSKOS_HOST is set then acquire an GCP account from Boskos.
-# Check out the account from Boskos and store the produced environment
-# variables in a temporary file.
-account_env_var_file="$(mktemp)"
-python hack/checkout_account.py 1>"${account_env_var_file}"
-checkout_account_status="${?}"
+if [ -n "${BOSKOS_HOST:-}" ]; then
+  # Check out the account from Boskos and store the produced environment
+  # variables in a temporary file.
+  account_env_var_file="$(mktemp)"
+  python hack/checkout_account.py 1>"${account_env_var_file}"
+  checkout_account_status="${?}"
 
-# If the checkout process was a success then load the account's
-# environment variables into this process.
-# shellcheck disable=SC1090
-[ "${checkout_account_status}" = "0" ] && . "${account_env_var_file}"
+  # If the checkout process was a success then load the account's
+  # environment variables into this process.
+  # shellcheck disable=SC1090
+  [ "${checkout_account_status}" = "0" ] && . "${account_env_var_file}"
 
-# Always remove the account environment variable file. It contains
-# sensitive information.
-rm -f "${account_env_var_file}"
+  # Always remove the account environment variable file. It contains
+  # sensitive information.
+  rm -f "${account_env_var_file}"
 
-if [ ! "${checkout_account_status}" = "0" ]; then
-  echo "error getting account from boskos" 1>&2
-  exit "${checkout_account_status}"
+  if [ ! "${checkout_account_status}" = "0" ]; then
+    echo "error getting account from boskos" 1>&2
+    exit "${checkout_account_status}"
+  fi
+
+  # run the heart beat process to tell boskos that we are still
+  # using the checked out account periodically
+  python -u hack/heartbeat_account.py >> "$ARTIFACTS/logs/boskos.log" 2>&1 &
+  HEART_BEAT_PID=$(echo $!)
 fi
-
-# run the heart beat process to tell boskos that we are still
-# using the checked out account periodically
-ARTIFACTS="${ARTIFACTS:-${PWD}/_artifacts}"
-mkdir -p "$ARTIFACTS/logs/"
-python -u hack/heartbeat_account.py >> "$ARTIFACTS/logs/boskos.log" 2>&1 &
-HEART_BEAT_PID=$(echo $!)
 
 hack/ci/e2e-conformance.sh --verbose $*
 test_status="${?}"
